Update license to AGPLv3 and simplify documentation
Changed the project license from GNU GPLv3 to GNU AGPLv3 in the LICENSE and updated references in the README. The README installation section was simplified to refer to an external guide. Minor clarification was made in the changelog, and the EditUserModal Svelte component's modal size was changed from 'm' to 'ms'.